refactor(perl-corpus): split corpus parsing/query into SRP modules (#0000)
